+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 0.0.16 (upcoming)
 - repaire the update_raw function of StationND
 - change data source from regnie to hyras for precipitation regionalisation
+- add ability to get nearby ma value from rasters, up to 1km from the station
 ## Version 0.0.15
 - change append with pandas concat function. -> faster
 - don't import complete module on installation